pyproject.toml
[tool.poetry]
name = "course-project"
version = "0.1.0"
description = ""
authors = ["Vadim Burin <burinvad@mail.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.13"
pandas = "^2.2.3"
requests = "^2.32.3"
python-dotenv = "^1.0.1"
openpyxl = "^3.1.5"
pytest = "^8.3.4"


[tool.poetry.group.dev.dependencies]
pytest-cov = "^6.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
